Update readme and add a few more tests
